+++ samples/sample_c/module_sample/camera_emu/dji_media_file_manage/dji_media_file_jpg.h
@@ -49,7 +49,7 @@ T_DjiReturnCode DjiMediaFile_GetAttrFunc_JPG(struct _DjiMediaFile *mediaFileHand
                                              T_DjiCameraMediaFileAttr *mediaFileAttr);
 
 T_DjiReturnCode DjiMediaFile_GetDataOrigin_JPG(struct _DjiMediaFile *mediaFileHandle, uint32_t offset, uint16_t len,
-                                               uint8_t *data, uint16_t *realLen);
+                                               uint8_t *data, uint32_t *realLen);
 T_DjiReturnCode DjiMediaFile_GetFileSizeOrigin_JPG(struct _DjiMediaFile *mediaFileHandle, uint32_t *fileSize);
 
 T_DjiReturnCode DjiMediaFile_CreateThumbNail_JPG(struct _DjiMediaFile *mediaFileHandle);
